requirements.txt
mistune==2.0.0rc1
requests==2.25.1
click==7.1.2
rich==10.1.0
Pillow==8.2.0
pytest==6.2.4
pytest-cov==2.11.1
pytest-mock==3.6.1
tox==3.23.1
